fixed: date and time entered correctly
